backend/python_scripts/requirements.txt
google-genai>=0.8.0
openai>=1.0.0
haystack-ai>=2.6.0
qdrant-haystack>=0.1.0
sentence-transformers>=2.7.0
python-dotenv>=1.0.0
qdrant-client>=1.9.0
transformers>=4.40.0
torch>=2.0.0